[skip ci]Refactor class name
